Bug 2031365 - Prevent implicit conversion of size_t arguments when calling js_pod alloc functions r=baku,spidermonkey-reviewers,sfink
Differential Revision: https://phabricator.services.mozilla.com/D294038

UltraBlame original commit: 8e0d2cdf22f85c4f0a456593f876d2d7b9770d2d

+++ js/xpconnect/src/XPCJSRuntime.cpp
@@ -3003,7 +3003,7 @@ static nsresult ReadSourceFromFilename(JSContext* cx, const char* filename,
   
   
   
-  JS::UniqueChars buf(js_pod_malloc<char>(rawLen));
+  JS::UniqueChars buf(js_pod_malloc<char>(static_cast<size_t>(rawLen)));
   if (!buf) {
     return NS_ERROR_OUT_OF_MEMORY;
   }
